Allow sharing of hero images
Extract all hero image code into a single module
