+++ config/locales/en.json
@@ -64,20 +64,17 @@
 			"title": "Big Lottery Fund",
 			"description": "The Big Lottery Fund gives grants to organisations in the UK to help improve their communities. The money awarded comes from the UK National Lottery.",
 			"home": "Home",
-			"intro": {
-				"long": "The Big Lottery Fund distributes <a href=\"/data\">over £600m</a> a year to communities across the UK, raised by players of the <a href=\"https://www.national-lottery.co.uk/\">National Lottery</a>",
-				"short": "The Big Lottery Fund distributes money raised by players of the <a href=\"https://www.national-lottery.co.uk/\">National Lottery</a> to communities across the UK."
-			},
-			"callsToAction": {
-				"under10k": {
-					"short": "Apply for funding under £10,000",
-					"long": "Learn how to apply for funding under £10,000"
-				},
-				"over10k": {
-					"short": "Apply for funding over £10,000",
-					"long": "Learn how to apply for funding over £10,000"
-				}
-			},
+			"hero": {
+                "intro": "The Big Lottery Fund distributes <a href=\"/data\">over £600m</a> a year to communities across the UK, raised by players of the <a href=\"https://www.national-lottery.co.uk/\">National Lottery</a>",
+                "under10k": {
+                    "short": "Apply for funding under £10,000",
+                    "long": "Learn how to apply for funding under £10,000"
+                },
+                "over10k": {
+                    "short": "Apply for funding over £10,000",
+                    "long": "Learn how to apply for funding over £10,000"
+                }
+            },
 			"caseStudies": [
 				{
 					"quoteText": "I am helping make sure that young people get the opportunities they want in terms of education.",
