+++ modules/utilities.js
@@ -1,6 +1,4 @@
 'use strict';
-const { has } = require('lodash');
-const assets = require('./assets');
 
 // rewrite text like '/welsh/' => '/welsh'
 const stripTrailingSlashes = str => {
@@ -28,26 +26,7 @@ const parseValueFromString = str => {
     return upper || str;
 };
 
-const createHeroImage = opts => {
-    if (!['small', 'medium', 'large'].every(x => has(opts, x))) {
-        throw new Error('Must pass a small, medium, and large image');
-    }
-
-    if (!has(opts, 'default')) {
-        throw new Error('Must define a default image with opts.default');
-    }
-
-    return {
-        small: assets.getImagePath(opts.small),
-        medium: assets.getImagePath(opts.medium),
-        large: assets.getImagePath(opts.large),
-        default: assets.getImagePath(opts.default),
-        caption: opts.caption || ''
-    };
-};
-
 module.exports = {
     stripTrailingSlashes,
-    parseValueFromString,
-    createHeroImage
+    parseValueFromString
 };
