+++ views/pages/toplevel/styleguide.njk
@@ -1,6 +1,5 @@
 {% extends "../../layouts/styleguide.njk" %}
 
-{% import "../../components/hero.njk" as hero %}
 {% from "../../components/caseStudy.njk" import caseStudy %}
 {% from "../../components/data.njk" import statsGrid %}
 {% from "../../components/hero.njk" import homepageSuperHero, sectionHero, miniatureHero %}
